Fix passing state containing regex with backslashes down to the webview
We used to pass `state` to the webview by converting it to a JS expression string to be evaluated.

This led to special characters (e.g. backslashes) not being escaped properly in regular expressions (they need to be escaped twice for this to work) which, in turn, prevented the web app from parsing this JSON string back into an object.

The workaround is to pass the same stringified state but encoded it as base64 so we don't have to worry about escaping special characters